+++ projects/demo/src/main/java/org/openmuc/jeebus/ship/examples/ExampleClient.java
@@ -16,6 +16,7 @@ import org.openmuc.jeebus.ship.node.ShipConfig;
 import org.openmuc.jeebus.ship.api.ShipConnectionInterface;
 
 import java.io.IOException;
+import java.net.URI;
 import java.nio.charset.StandardCharsets;
 
 public class ExampleClient {
@@ -49,7 +50,8 @@ public class ExampleClient {
 
         // replace String parameter with server IP as needed
         ShipConnectionInterface shipConnInterface = ship.openConnection(
-            "localhost:2001");
+            URI.create("wss://localhost:2001/ship/")
+        );
 
         /*
          The communication partner can also be authenticated after opening the