+++ projects/demo/src/main/java/org/openmuc/jeebus/ship/examples/MainTest.java
@@ -16,6 +16,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.net.URI;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.time.Instant;
@@ -48,12 +49,12 @@ public class MainTest {
             }
 
             @Override
-            public void serviceAdded(String serviceId, String ski) {
+            public void serviceAdded(ShipService service) {
 
             }
 
             @Override
-            public void serviceRemoved(String serviceId) {
+            public void serviceRemoved(ShipService service) {
 
             }
 
@@ -82,12 +83,12 @@ public class MainTest {
             }
 
             @Override
-            public void serviceAdded(String serviceId, String ski) {
+            public void serviceAdded(ShipService service) {
 
             }
 
             @Override
-            public void serviceRemoved(String serviceId) {
+            public void serviceRemoved(ShipService service) {
 
             }
 
@@ -125,7 +126,7 @@ public class MainTest {
         LOGGER.info("Opening connection...");
 
         ShipConnectionInterface shipConnInterface = ship2.openConnection(
-            "127.0.0.1:8080"
+            URI.create("wss://127.0.0.1:8080/ship/")
         );
 
         Instant connectionReady = Instant.now();