+++ projects/ship/src/main/java/org/openmuc/jeebus/ship/node/ShipNodeImpl.java
@@ -28,14 +28,13 @@ import org.slf4j.LoggerFactory;
 import javax.net.ssl.SSLException;
 import java.net.InetSocketAddress;
 import java.net.URI;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static org.openmuc.jeebus.ship.node.KeyManagement.encodeSkiAsString;
 public class ShipNodeImpl {
@@ -131,9 +130,17 @@ public class ShipNodeImpl {
                 getOwnSki(),
                 server.getBoundSocketAddresses()
             );
-
         }
+    }
 
+    public Set<WebSocketHandler> getAllWebSocketHandlers() {
+        return Stream.concat(
+            servers.stream()
+                .map(ShipServer::getHandlers)
+                .flatMap(Collection::stream),
+            clients.stream()
+                .map(ShipClient::getHandler)
+        ).collect(Collectors.toSet());
     }
 
     public void closeDoubleConns(WebSocketHandler current) {